ci(engine): add ruff check as CI gate + fix pre-existing lint errors (closes #49)
engine-build.yml test job에 ruff check 스텝 추가. 게이트화에 앞서
코드베이스의 pre-existing lint 에러 9건 정리:

- F401: main.py / test_solvers_base.py unused import 제거
- F841: test_solvers_{gurobi,ortools}.py unused `result` 변수 제거
- B017: test_graph.py blind `Exception` → `FrozenInstanceError`
- UP035: main.py AsyncIterator를 collections.abc에서 import
- RUF100: main.py dead `# noqa: PLW0603` 제거
- UP042: models.py Status `(str, Enum)` → `StrEnum`
  (.value 접근·Pydantic 직렬화 동일, test_api.py로 계약 무회귀 확인)
- RUF046: gurobi.py `int(round(...))` → `round(...)`
- RUF003: test_graph.py 주석의 모호한 `×` → `*`

+++ engine/src/solvers/gurobi.py
@@ -251,8 +251,8 @@ class GurobiSolver(BaseSolver):
 
         # ── y[d] 결정변수 노출 ───────────────────────────────
         self._last_y_values = {
-            **{c: int(round(y_city[c].X)) for c in graph.internal_cities},
-            **{h: int(round(y_hub[h].X)) for h in graph.hubs},
+            **{c: round(y_city[c].X) for c in graph.internal_cities},
+            **{h: round(y_hub[h].X) for h in graph.hubs},
         }
 
         # 체류시간 추가 (deadline 제약과 의미 일치)
